Add a UserMapping to store CAS/SAML provider ids
In our opinion, the best practice when it comes to third party
authentication, is to create models that store mappings between
the third party ids and our local users.

This lightens requirements on consumers' user models and is more
flexible when it comes to combining multiple authentication methods
and/or providers.