Use our UserMapping in the authentication backend
Now that we have a model that enables linking Users with their
CAS/SAML provider ids, we can use it to find existing users as they
log in or create them along with their mapping if they do not have
an account yet.

We also removed some code that handled field dependencies on the actual
User model as we don't want to rely on those anymore.