Update setup.py & setup.cfg to make a new package
As we make breaking changes, our immediate goal is not to merge
our changes back into the django-cas-ng package but to create a
new package that fulfills our own needs, and then start a discussion
about porting our changed to the original codebase.

Therefore, we need to change setupfiles to make our own version of
the package with a different name suffix.