Add a new setting to force logout redirect URL
Logging users out of the Django app without redirecting them to
a view accessible to anonymous users is pretty pointless: it
eithers loops them back through login instantly if we did not force
log out of the CAS/SAML provider, or just sends them back to their
provider to log in otherwise.

Both of these scenarios are user-hostile and are bugs in our opinion.
We'd rather force a redirect to a given URL that maps to a view
anonymous users can access. This way, users can see that they were
logged out and be on their way (or log in again / log in with another
account).